refactor : 서버 3차 정리
