--- src/main/java/com/example/team258/common/config/SemaphoreConfig.java
@@ -1,14 +1,0 @@
-package com.example.team258.common.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import java.util.concurrent.Semaphore;
-
-@Configuration
-public class SemaphoreConfig {
-    @Bean
-    public Semaphore binarySemaphore() {
-        return new Semaphore(1, true);
-    }
-}
