+++ src/main/java/com/example/team258/common/entity/User.java
@@ -1,15 +1,14 @@
 package com.example.team258.common.entity;
 
 
-import com.example.team258.domain.donation.entity.BookApplyDonation;
-import com.example.team258.domain.user.entity.BookRent;
-import com.example.team258.domain.user.entity.BookReservation;
+import com.example.team258.kafka.entity.BookApplyDonation;
+import com.example.team258.kafka.entity.BookRent;
+import com.example.team258.kafka.entity.BookReservation;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.annotations.BatchSize;
 
 import java.util.ArrayList;
 import java.util.List;
