+++ src/main/java/com/example/team258/domain/donation/dto/BookDonationEventResponseDto.java
@@ -1,7 +1,7 @@
 package com.example.team258.domain.donation.dto;
 
 import com.example.team258.common.dto.BookResponseDto;
-import com.example.team258.domain.donation.entity.BookDonationEvent;
+import com.example.team258.kafka.entity.BookDonationEvent;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
