+++ src/main/java/com/example/team258/domain/donation/dto/BookDonationEventResponseDtoV3.java
@@ -1,6 +1,6 @@
 package com.example.team258.domain.donation.dto;
 import com.example.team258.common.dto.BookResponseDto;
-import com.example.team258.domain.donation.entity.BookDonationEvent;
+import com.example.team258.kafka.entity.BookDonationEvent;
 import lombok.Data;
 
 import java.time.LocalDateTime;
