--- src/main/java/com/example/team258/domain/member/dto/LoginRequestDto.java
@@ -1,9 +1,0 @@
-package com.example.team258.domain.member.dto;
-
-import lombok.Getter;
-
-@Getter
-public class LoginRequestDto {
-    private String username;
-    private String password;
-}
