--- src/main/java/com/example/team258/kafka/dto/UserResponseKafkaDto.java
@@ -1,21 +1,0 @@
-package com.example.team258.kafka.dto;
-
-import com.example.team258.domain.member.dto.UserResponseDto;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-public class UserResponseKafkaDto {
-    private List<UserResponseDto> userResponseDtos;
-    private int page;
-    private int totalPages;
-    private String correlationId;
-    public UserResponseKafkaDto(List<UserResponseDto> userResponseDtos, int page, int totalPages) {
-        this.userResponseDtos = userResponseDtos;
-        this.page = page;
-        this.totalPages = totalPages;
-    }
-}
