+++ src/main/java/com/example/team258/kafka/entity/BookRent.java
@@ -1,4 +1,4 @@
-package com.example.team258.domain.user.entity;
+package com.example.team258.kafka.entity;
 
 import com.example.team258.common.entity.Book;
 import jakarta.persistence.*;
