+++ src/main/java/com/example/team258/kafka/entity/BookReservation.java
@@ -1,4 +1,4 @@
-package com.example.team258.domain.user.entity;
+package com.example.team258.kafka.entity;
 
 import com.example.team258.common.entity.Book;
 import com.example.team258.common.entity.User;
