+++ src/main/java/com/example/team258/kafka/repository/BookRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.example.team258.common.repository;
+package com.example.team258.kafka.repository;
 
 import com.example.team258.common.entity.BookCategory;
 import com.example.team258.common.entity.Book;
