+++ src/test/java/BatchSizeTest.java
@@ -1,6 +1,6 @@
 //import com.example.team258.common.entity.User;
-//import com.example.team258.common.repository.UserRepository;
-//import com.example.team258.domain.donation.entity.BookApplyDonation;
+//import com.example.team258.kafka.repository.UserRepository;
+//import com.example.team258.kafka.entity.BookApplyDonation;
 //import org.junit.jupiter.api.Test;
 //import org.springframework.beans.factory.annotation.Autowired;
 //import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
